add PrivateKeyToKeypair; working on unlock
